Refactor, remove cruft.
